+++ internal/handlers/http/me.go
@@ -1,33 +1,12 @@
 package httpHandlers
 
 import (
-	"fmt"
-
 	"github.com/gofiber/fiber/v2"
-	g "github.com/maktoobgar/go_template/internal/global"
-	user_service "github.com/maktoobgar/go_template/internal/services/users"
+	"github.com/maktoobgar/go_template/internal/models"
 )
 
 func Me(c *fiber.Ctx) error {
-	uService := user_service.New()
-
-	session, err := g.Session.Get(c)
-	if err != nil {
-		return err
-	}
-
-	if session.Fresh() {
-		session.Destroy()
-		c.Redirect("SignIn")
-		return nil
-	}
-
-	userID := session.Get(session.ID()).(int)
-	user, err := uService.GetUserByID(fmt.Sprint(userID))
-	if err != nil {
-		return err
-	}
-
+	user := c.Locals("user").(models.User)
 	c.JSON(user.Clean())
 	return nil
 }