+++ internal/handlers/http/signin.go
@@ -3,18 +3,20 @@ package httpHandlers
 import (
 	"github.com/gofiber/fiber/v2"
 	g "github.com/maktoobgar/go_template/internal/global"
+	"github.com/maktoobgar/go_template/internal/handlers/utils"
 	auth_service "github.com/maktoobgar/go_template/internal/services/auth"
+	"github.com/maktoobgar/go_template/pkg/errors"
 )
 
 type signInRequest struct {
-	Username string `json:"username" xml:"username" form:"username"`
-	Password string `json:"password" xml:"password" form:"password"`
+	Username string `json:"username" xml:"username" form:"username" required:"true"`
+	Password string `json:"password" xml:"password" form:"password" required:"true"`
 }
 
 func SignIn(c *fiber.Ctx) error {
 	req := &signInRequest{}
-	if err := c.BodyParser(req); err != nil {
-		return err
+	if err := c.BodyParser(req); err != nil || !utils.Required(req) {
+		return errors.New(errors.InvalidStatus, "not all required fields provided")
 	}
 
 	auth := auth_service.New()