+++ internal/routers/http.go
@@ -9,7 +9,12 @@ import (
 func Http(app *fiber.App) {
 	// /api
 	apiGroup := app.Group("/api", middleware.Useless)
-	apiGroup.Get("me", httpHandler.Me).Name("Me")
+
+	// /api/me
+	meGroup := apiGroup.Group("/me", middleware.SessionAuth)
+	meGroup.Get("/", httpHandler.Me).Name("Me")
+
+	// /api/:name?
 	apiGroup.Get(":name?", httpHandler.Hi).Name("Hi")
 
 	// /api/auth